+++ Source/VMware.vSphereDSC/VMware.vSphereDSC.psm1
@@ -24,6 +24,14 @@ enum AcceptanceLevel {
     CommunitySupported
 }
 
+enum ChapType {
+    Prohibited
+    Discouraged
+    Preferred
+    Required
+    Unset
+}
+
 enum DomainAction {
     Join
     Leave
@@ -63,6 +71,11 @@ enum FolderType {
 enum GraphicsType {
     Shared
     SharedDirect
+}
+
+enum IScsiHbaTargetType {
+    Static
+    Send
 }
 
 enum LinkDiscoveryProtocolOperation {
@@ -1260,6 +1273,171 @@ class VMHostGraphicsBaseDSC : VMHostBaseDSC {
     #>
     [string] ConvertEnumValueToServerValue($enumValue) {
         return $enumValue.ToString().Substring(0, 1).ToLower() + $enumValue.ToString().Substring(1)
+    }
+}
+
+class VMHostIScsiHbaBaseDSC : VMHostEntityBaseDSC {
+    <#
+    .DESCRIPTION
+
+    Specifies the type of the CHAP (Challenge Handshake Authentication Protocol).
+    #>
+    [DscProperty()]
+    [ChapType] $ChapType = [ChapType]::Unset
+
+    <#
+    .DESCRIPTION
+
+    Specifies the CHAP authentication name.
+    #>
+    [DscProperty()]
+    [string] $ChapName
+
+    <#
+    .DESCRIPTION
+
+    Specifies the CHAP authentication password.
+    #>
+    [DscProperty()]
+    [string] $ChapPassword
+
+    <#
+    .DESCRIPTION
+
+    Indicates that Mutual CHAP is enabled.
+    #>
+    [DscProperty()]
+    [nullable[bool]] $MutualChapEnabled
+
+    <#
+    .DESCRIPTION
+
+    Specifies the Mutual CHAP authentication name.
+    #>
+    [DscProperty()]
+    [string] $MutualChapName
+
+    <#
+    .DESCRIPTION
+
+    Specifies the Mutual CHAP authentication password.
+    #>
+    [DscProperty()]
+    [string] $MutualChapPassword
+
+    <#
+    .DESCRIPTION
+
+    Specifies whether to change the password for CHAP, Mutual CHAP or both. When the property is not specified or its value is $false, it is ignored.
+    If the property is $true the passwords for CHAP and Mutual CHAP are changed to their desired values.
+    #>
+    [DscProperty()]
+    [nullable[bool]] $Force
+
+    hidden [string] $IScsiDeviceType = 'iSCSI'
+
+    hidden [string] $CouldNotRetrieveIScsiHbaMessage = "Could not retrieve iSCSI Host Bus Adapter {0} from VMHost {1}. For more information: {2}"
+
+    <#
+    .DESCRIPTION
+
+    Retrieves the iSCSI Host Bus Adapter with the specified name from the specified VMHost if it exists.
+    #>
+    [PSObject] GetIScsiHba($iScsiHbaName) {
+        try {
+            $iScsiHba = Get-VMHostHba -Server $this.Connection -VMHost $this.VMHost -Device $iScsiHbaName -Type $this.IScsiDeviceType -ErrorAction Stop -Verbose:$false
+            return $iScsiHba
+        }
+        catch {
+            throw ($this.CouldNotRetrieveIScsiHbaMessage -f $iScsiHbaName, $this.VMHost.Name, $_.Exception.Message)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Checks if the CHAP settings should be modified based on the current authentication properties.
+    #>
+    [bool] ShouldModifyCHAPSettings($authenticationProperties, $inheritChap, $inheritMutualChap) {
+        $shouldModifyCHAPSettings = @()
+
+        $shouldModifyCHAPSettings += ($null -ne $inheritChap -and $inheritChap -ne $authenticationProperties.ChapInherited)
+        $shouldModifyCHAPSettings += ($this.ChapType -ne [ChapType]::Unset -and $this.ChapType.ToString() -ne $authenticationProperties.ChapType.ToString())
+        $shouldModifyCHAPSettings += ($null -ne $inheritMutualChap -and $inheritMutualChap -ne $authenticationProperties.MutualChapInherited)
+        $shouldModifyCHAPSettings += ($null -ne $this.MutualChapEnabled -and $this.MutualChapEnabled -ne $authenticationProperties.MutualChapEnabled)
+
+        # Force should determine the Desired State only when it is $true.
+        $shouldModifyCHAPSettings += ($null -ne $this.Force -and $this.Force)
+
+        # CHAP and Mutual CHAP names should be ignored when determining the Desired State when CHAP type is 'Prohibited'.
+        if ($this.ChapType -ne [ChapType]::Prohibited) {
+            $shouldModifyCHAPSettings += (![string]::IsNullOrEmpty($this.ChapName) -and $this.ChapName -ne [string] $authenticationProperties.ChapName)
+            $shouldModifyCHAPSettings += (![string]::IsNullOrEmpty($this.MutualChapName) -and $this.MutualChapName -ne [string] $authenticationProperties.MutualChapName)
+        }
+
+        return ($shouldModifyCHAPSettings -Contains $true)
+    }
+
+    <#
+    .DESCRIPTION
+
+    Checks if the CHAP settings should be modified based on the current authentication properties.
+    #>
+    [bool] ShouldModifyCHAPSettings($authenticationProperties) {
+        return $this.ShouldModifyCHAPSettings($authenticationProperties, $null, $null)
+    }
+
+    <#
+    .DESCRIPTION
+
+    Populates the cmdlet parameters with the CHAP settings based on the following criteria:
+    1. CHAP settings can only be passed to the cmdlet if the 'InheritChap' option is not passed or it is passed with a '$false' value.
+    2. Mutual CHAP settings can only be passed to the cmdlet if the 'InheritMutualChap' option is not passed or it is passed with a '$false' value.
+    3. CHAP name and CHAP password can be passed to the cmdlet if the CHAP type is not 'Prohibited'.
+    4. Mutual CHAP settings can only be passed to the cmdlet if the CHAP type is 'Required'.
+    5. Mutual CHAP name and Mutual CHAP password can be passed to the cmdlet if Mutual CHAP enabled is not passed
+       or if it is passed with a '$true' value.
+    #>
+    [void] PopulateCmdletParametersWithCHAPSettings($cmdletParams, $inheritChap, $inheritMutualChap) {
+        if ($null -ne $inheritChap -and $inheritChap) {
+            $cmdletParams.InheritChap = $inheritChap
+        }
+        else {
+            # When 'InheritChap' is $false, it can be passed to the cmdlet only if CHAP type is not 'Prohibited'.
+            if ($null -ne $inheritChap -and $this.ChapType -ne [ChapType]::Prohibited) { $cmdletParams.InheritChap = $inheritChap }
+            if ($this.ChapType -ne [ChapType]::Unset) { $cmdletParams.ChapType = $this.ChapType.ToString() }
+
+            if ($this.ChapType -ne [ChapType]::Prohibited) {
+                if (![string]::IsNullOrEmpty($this.ChapName)) { $cmdletParams.ChapName = $this.ChapName }
+                if (![string]::IsNullOrEmpty($this.ChapPassword)) { $cmdletParams.ChapPassword = $this.ChapPassword }
+            }
+        }
+
+        if ($null -ne $inheritMutualChap -and $inheritMutualChap) {
+            $cmdletParams.InheritMutualChap = $inheritMutualChap
+        }
+        else {
+            # When 'InheritMutualChap' is $false, it can be passed to the cmdlet only if CHAP type is not 'Prohibited'.
+            if ($null -ne $inheritMutualChap -and $this.ChapType -ne [ChapType]::Prohibited) { $cmdletParams.InheritMutualChap = $inheritMutualChap }
+
+            if ($this.ChapType -eq [ChapType]::Required) {
+                if ($null -ne $this.MutualChapEnabled) { $cmdletParams.MutualChapEnabled = $this.MutualChapEnabled }
+
+                if ($null -eq $this.MutualChapEnabled -or $this.MutualChapEnabled) {
+                    if (![string]::IsNullOrEmpty($this.MutualChapName)) { $cmdletParams.MutualChapName = $this.MutualChapName }
+                    if (![string]::IsNullOrEmpty($this.MutualChapPassword)) { $cmdletParams.MutualChapPassword = $this.MutualChapPassword }
+                }
+            }
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Populates the cmdlet parameters with the CHAP settings.
+    #>
+    [void] PopulateCmdletParametersWithCHAPSettings($cmdletParams) {
+        $this.PopulateCmdletParametersWithCHAPSettings($cmdletParams, $null, $null)
     }
 }
 
@@ -6517,6 +6695,419 @@ class VMHostFirewallRuleset : VMHostEntityBaseDSC {
         $vmHostFirewallRulesetAllowedHosts = $vmHostFirewallRuleset.ExtensionData.AllowedHosts
         $result.AllIP = $vmHostFirewallRulesetAllowedHosts.AllIp
         $result.IPAddresses = $vmHostFirewallRulesetAllowedHosts.IpAddress + $this.ConvertHostFirewallRulesetIpNetworksToIPNetworks($vmHostFirewallRulesetAllowedHosts.IpNetwork)
+    }
+}
+
+[DscResource()]
+class VMHostIScsiHba : VMHostIScsiHbaBaseDSC {
+    <#
+    .DESCRIPTION
+
+    Specifies the name of the iSCSI Host Bus Adapter.
+    #>
+    [DscProperty(Key)]
+    [string] $Name
+
+    hidden [string] $ConfigureIScsiHbaChapMessage = "Configuring CHAP settings of iSCSI Host Bus Adapter {0} from VMHost {1}."
+
+    hidden [string] $CouldNotConfigureIScsiHbaChapMessage = "Could not configure CHAP settings of iSCSI Host Bus Adapter {0} from VMHost {1}. For more information: {2}"
+
+    [void] Set() {
+        try {
+            Write-VerboseLog -Message $this.SetMethodStartMessage -Arguments @($this.DscResourceName)
+            $this.ConnectVIServer()
+            $this.RetrieveVMHost()
+
+            $iScsiHba = $this.GetIScsiHba($this.Name)
+
+            $this.ConfigureIScsiHbaChap($iScsiHba)
+        }
+        finally {
+            $this.DisconnectVIServer()
+            Write-VerboseLog -Message $this.SetMethodEndMessage -Arguments @($this.DscResourceName)
+        }
+    }
+
+    [bool] Test() {
+        try {
+            Write-VerboseLog -Message $this.TestMethodStartMessage -Arguments @($this.DscResourceName)
+            $this.ConnectVIServer()
+            $this.RetrieveVMHost()
+
+            $iScsiHba = $this.GetIScsiHba($this.Name)
+
+            $result = !$this.ShouldModifyCHAPSettings($iScsiHba.AuthenticationProperties)
+
+            $this.WriteDscResourceState($result)
+
+            return $result
+        }
+        finally {
+            $this.DisconnectVIServer()
+            Write-VerboseLog -Message $this.TestMethodEndMessage -Arguments @($this.DscResourceName)
+        }
+    }
+
+    [VMHostIScsiHba] Get() {
+        try {
+            Write-VerboseLog -Message $this.GetMethodStartMessage -Arguments @($this.DscResourceName)
+            $result = [VMHostIScsiHba]::new()
+
+            $this.ConnectVIServer()
+            $this.RetrieveVMHost()
+
+            $iScsiHba = $this.GetIScsiHba($this.Name)
+
+            $this.PopulateResult($result, $iScsiHba)
+
+            return $result
+        }
+        finally {
+            $this.DisconnectVIServer()
+            Write-VerboseLog -Message $this.GetMethodEndMessage -Arguments @($this.DscResourceName)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Configures the CHAP properties of the specified iSCSI Host Bus Adapter.
+    #>
+    [void] ConfigureIScsiHbaChap($iScsiHba) {
+        $setVMHostHbaParams = @{
+            IScsiHba = $iScsiHba
+            Confirm = $false
+            ErrorAction = 'Stop'
+            Verbose = $false
+        }
+
+        $this.PopulateCmdletParametersWithCHAPSettings($setVMHostHbaParams)
+
+        try {
+            Write-VerboseLog -Message $this.ConfigureIScsiHbaChapMessage -Arguments @($iScsiHba.Device, $this.VMHost.Name)
+            Set-VMHostHba @setVMHostHbaParams
+        }
+        catch {
+            throw ($this.CouldNotConfigureIScsiHbaChapMessage -f $iScsiHba.Device, $this.VMHost.Name, $_.Exception.Message)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Populates the result returned from the Get method.
+    #>
+    [void] PopulateResult($result, $iScsiHba) {
+        $result.Server = $this.Connection.Name
+        $result.VMHostName = $this.VMHost.Name
+        $result.Name = $iScsiHba.Device
+        $result.ChapType = $iScsiHba.AuthenticationProperties.ChapType.ToString()
+        $result.ChapName = [string] $iScsiHba.AuthenticationProperties.ChapName
+        $result.MutualChapEnabled = $iScsiHba.AuthenticationProperties.MutualChapEnabled
+        $result.MutualChapName = [string] $iScsiHba.AuthenticationProperties.MutualChapName
+        $result.Force = $this.Force
+    }
+}
+
+[DscResource()]
+class VMHostIScsiHbaTarget : VMHostIScsiHbaBaseDSC {
+    <#
+    .DESCRIPTION
+
+    Specifies the address of the iSCSI Host Bus Adapter target.
+    #>
+    [DscProperty(Key)]
+    [string] $Address
+
+    <#
+    .DESCRIPTION
+
+    Specifies the TCP port of the iSCSI Host Bus Adapter target.
+    #>
+    [DscProperty(Key)]
+    [int] $Port
+
+    <#
+    .DESCRIPTION
+
+    Specifies the name of the iSCSI Host Bus Adapter of the iSCSI Host Bus Adapter target.
+    #>
+    [DscProperty(Key)]
+    [string] $IScsiHbaName
+
+    <#
+    .DESCRIPTION
+
+    Specifies the type of the iSCSI Host Bus Adapter target.
+    #>
+    [DscProperty(Key)]
+    [IScsiHbaTargetType] $TargetType
+
+    <#
+    .DESCRIPTION
+
+    Specifies whether the iSCSI Host Bus Adapter target should be present or absent.
+    #>
+    [DscProperty(Mandatory)]
+    [Ensure] $Ensure
+
+    <#
+    .DESCRIPTION
+
+    Specifies the iSCSI name of the iSCSI Host Bus Adapter target. It is required for Static iSCSI Host Bus Adapter targets.
+    #>
+    [DscProperty()]
+    [string] $IScsiName
+
+    <#
+    .DESCRIPTION
+
+    Indicates that the CHAP setting is inherited from the iSCSI Host Bus Adapter.
+    #>
+    [DscProperty()]
+    [nullable[bool]] $InheritChap
+
+    <#
+    .DESCRIPTION
+
+    Indicates that the Mutual CHAP setting is inherited from the iSCSI Host Bus Adapter.
+    #>
+    [DscProperty()]
+    [nullable[bool]] $InheritMutualChap
+
+    <#
+    .DESCRIPTION
+
+    Specifies the <Address>:<Port> that uniquely identifies an iSCSI Host Bus Adapter target.
+    #>
+    hidden [string] $IPEndPoint
+
+    hidden [string] $CreateIScsiHbaTargetMessage = "Creating iSCSI Host Bus Adapter target with IP address {0} on iSCSI Host Bus Adapter device {1}."
+    hidden [string] $ModifyIScsiHbaTargetMessage = "Modifying CHAP settings of iSCSI Host Bus Adapter target with IP address {0} on iSCSI Host Bus Adapter device {1}."
+    hidden [string] $RemoveIScsiHbaTargetMessage = "Removing iSCSI Host Bus Adapter target with IP address {0} from iSCSI Host Bus Adapter device {1}."
+
+    hidden [string] $CouldNotCreateIScsiHbaTargetMessage = "Could not create iSCSI Host Bus Adapter target with IP address {0} on iSCSI Host Bus Adapter device {1}. For more information: {2}"
+    hidden [string] $CouldNotModifyIScsiHbaTargetMessage = "Could not modify CHAP settings of iSCSI Host Bus Adapter target with IP address {0} on iSCSI Host Bus Adapter device {1}. For more information: {2}"
+    hidden [string] $CouldNotRemoveIScsiHbaTargetMessage = "Could not remove iSCSI Host Bus Adapter target with IP address {0} from iSCSI Host Bus Adapter device {1}. For more information: {2}"
+
+    [void] Set() {
+        try {
+            Write-VerboseLog -Message $this.SetMethodStartMessage -Arguments @($this.DscResourceName)
+            $this.ConnectVIServer()
+            $this.RetrieveVMHost()
+            $this.IPEndPoint = $this.Address + ':' + $this.Port.ToString()
+
+            $iScsiHba = $this.GetIScsiHba($this.IScsiHbaName)
+            $iScsiHbaTarget = $this.GetIScsiHbaTarget($iScsiHba)
+
+            if ($this.Ensure -eq [Ensure]::Present) {
+                if ($null -eq $iScsiHbaTarget) {
+                    $this.NewIScsiHbaTarget($iScsiHba)
+                }
+                else {
+                    $this.ModifyIScsiHbaTarget($iScsiHbaTarget)
+                }
+            }
+            else {
+                if ($null -ne $iScsiHbaTarget) {
+                    $this.RemoveIScsiHbaTarget($iScsiHbaTarget)
+                }
+            }
+        }
+        finally {
+            $this.DisconnectVIServer()
+            Write-VerboseLog -Message $this.SetMethodEndMessage -Arguments @($this.DscResourceName)
+        }
+    }
+
+    [bool] Test() {
+        try {
+            Write-VerboseLog -Message $this.TestMethodStartMessage -Arguments @($this.DscResourceName)
+            $this.ConnectVIServer()
+            $this.RetrieveVMHost()
+            $this.IPEndPoint = $this.Address + ':' + $this.Port.ToString()
+
+            $iScsiHba = $this.GetIScsiHba($this.IScsiHbaName)
+            $iScsiHbaTarget = $this.GetIScsiHbaTarget($iScsiHba)
+
+            $result = $null
+
+            if ($this.Ensure -eq [Ensure]::Present) {
+                if ($null -eq $iScsiHbaTarget) {
+                    $result = $false
+                }
+                else {
+                    $result = !$this.ShouldModifyCHAPSettings($iScsiHbaTarget.AuthenticationProperties, $this.InheritChap, $this.InheritMutualChap)
+                }
+            }
+            else {
+                $result = ($null -eq $iScsiHbaTarget)
+            }
+
+            $this.WriteDscResourceState($result)
+
+            return $result
+        }
+        finally {
+            $this.DisconnectVIServer()
+            Write-VerboseLog -Message $this.TestMethodEndMessage -Arguments @($this.DscResourceName)
+        }
+    }
+
+    [VMHostIScsiHbaTarget] Get() {
+        try {
+            Write-VerboseLog -Message $this.GetMethodStartMessage -Arguments @($this.DscResourceName)
+            $result = [VMHostIScsiHbaTarget]::new()
+
+            $this.ConnectVIServer()
+            $this.RetrieveVMHost()
+            $this.IPEndPoint = $this.Address + ':' + $this.Port.ToString()
+
+            $iScsiHba = $this.GetIScsiHba($this.IScsiHbaName)
+            $iScsiHbaTarget = $this.GetIScsiHbaTarget($iScsiHba)
+
+            $this.PopulateResult($result, $iScsiHbaTarget)
+
+            return $result
+        }
+        finally {
+            $this.DisconnectVIServer()
+            Write-VerboseLog -Message $this.GetMethodEndMessage -Arguments @($this.DscResourceName)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Retrieves the iSCSI Host Bus Adapter target with the specified IPEndPoint for the specified iSCSI Host Bus Adapter if it exists.
+    #>
+    [PSObject] GetIScsiHbaTarget($iScsiHba) {
+        $getIScsiHbaTargetParams = @{
+            Server = $this.Connection
+            IScsiHba = $iScsiHba
+            IPEndPoint = $this.IPEndPoint
+            Type = $this.TargetType.ToString()
+            ErrorAction = 'SilentlyContinue'
+            Verbose = $false
+        }
+
+        return Get-IScsiHbaTarget @getIScsiHbaTargetParams
+    }
+
+    <#
+    .DESCRIPTION
+
+    Creates a new iSCSI Host Bus Adapter target of the specified type with the specified address for the specified iSCSI Host Bus Adapter.
+    #>
+    [void] NewIScsiHbaTarget($iScsiHba) {
+        $newIScsiHbaTargetParams = @{
+            Server = $this.Connection
+            Address = $this.Address
+            Port = $this.Port
+            IScsiHba = $iScsiHba
+            Type = $this.TargetType.ToString()
+            Confirm = $false
+            ErrorAction = 'Stop'
+            Verbose = $false
+        }
+
+        if ($this.TargetType -eq [IScsiHbaTargetType]::Static) { $newIScsiHbaTargetParams.IScsiName = $this.IScsiName }
+        $this.PopulateCmdletParametersWithCHAPSettings($newIScsiHbaTargetParams, $this.InheritChap, $this.InheritMutualChap)
+
+        try {
+            Write-VerboseLog -Message $this.CreateIScsiHbaTargetMessage -Arguments @($this.IPEndPoint, $this.IScsiHbaName)
+            New-IScsiHbaTarget @newIScsiHbaTargetParams
+        }
+        catch {
+            throw ($this.CouldNotCreateIScsiHbaTargetMessage -f $this.IPEndPoint, $this.IScsiHbaName, $_.Exception.Message)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Modifies the CHAP settings of the specified iSCSI Host Bus Adapter target.
+    #>
+    [void] ModifyIScsiHbaTarget($iScsiHbaTarget) {
+        $setIScsiHbaTargetParams = @{
+            Server = $this.Connection
+            Target = $iScsiHbaTarget
+            Confirm = $false
+            ErrorAction = 'Stop'
+            Verbose = $false
+        }
+
+        $this.PopulateCmdletParametersWithCHAPSettings($setIScsiHbaTargetParams, $this.InheritChap, $this.InheritMutualChap)
+
+        try {
+            Write-VerboseLog -Message $this.ModifyIScsiHbaTargetMessage -Arguments @($this.IPEndPoint, $this.IScsiHbaName)
+            Set-IScsiHbaTarget @setIScsiHbaTargetParams
+        }
+        catch {
+            throw ($this.CouldNotModifyIScsiHbaTargetMessage -f $this.IPEndPoint, $this.IScsiHbaName, $_.Exception.Message)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Removes the specified iSCSI Host Bus Adapter target from its iSCSI Host Bus Adapter.
+    #>
+    [void] RemoveIScsiHbaTarget($iScsiHbaTarget) {
+        $removeIScsiHbaTargetParams = @{
+            Server = $this.Connection
+            Target = $iScsiHbaTarget
+            Confirm = $false
+            ErrorAction = 'Stop'
+            Verbose = $false
+        }
+
+        try {
+            Write-VerboseLog -Message $this.RemoveIScsiHbaTargetMessage -Arguments @($this.IPEndPoint, $this.IScsiHbaName)
+            Remove-IScsiHbaTarget @removeIScsiHbaTargetParams
+        }
+        catch {
+            throw ($this.CouldNotRemoveIScsiHbaTargetMessage -f $this.IPEndPoint, $this.IScsiHbaName, $_.Exception.Message)
+        }
+    }
+
+    <#
+    .DESCRIPTION
+
+    Populates the result returned from the Get method.
+    #>
+    [void] PopulateResult($result, $iScsiHbaTarget) {
+        $result.Server = $this.Connection.Name
+        $result.VMHostName = $this.VMHost.Name
+        $result.IScsiHbaName = $this.IScsiHbaName
+        $result.Force = $this.Force
+
+        if ($null -ne $iScsiHbaTarget) {
+            $result.Address = $iScsiHbaTarget.Address
+            $result.Port = $iScsiHbaTarget.Port
+            $result.TargetType = $iScsiHbaTarget.Type.ToString()
+            $result.IScsiName = $iScsiHbaTarget.IScsiName
+            $result.InheritChap = $iScsiHbaTarget.AuthenticationProperties.ChapInherited
+            $result.ChapType = $iScsiHbaTarget.AuthenticationProperties.ChapType.ToString()
+            $result.ChapName = [string] $iScsiHbaTarget.AuthenticationProperties.ChapName
+            $result.InheritMutualChap = $iScsiHbaTarget.AuthenticationProperties.MutualChapInherited
+            $result.MutualChapEnabled = $iScsiHbaTarget.AuthenticationProperties.MutualChapEnabled
+            $result.MutualChapName = [string] $iScsiHbaTarget.AuthenticationProperties.MutualChapName
+            $result.Ensure = [Ensure]::Present
+        }
+        else {
+            $result.Address = $this.Address
+            $result.Port = $this.Port
+            $result.TargetType = $this.TargetType
+            $result.IScsiName = $this.IScsiName
+            $result.InheritChap = $this.InheritChap
+            $result.ChapType = $this.ChapType
+            $result.ChapName = $this.ChapName
+            $result.InheritMutualChap = $this.InheritMutualChap
+            $result.MutualChapEnabled = $this.MutualChapEnabled
+            $result.MutualChapName = $this.MutualChapName
+            $result.Ensure = [Ensure]::Absent
+        }
     }
 }